refactor: convert lint_templates.py to pytest+hypothesis test
