+++ .github/workflows/ci.yml
@@ -100,7 +100,7 @@ jobs:
     - name: Test project generation and workflow
       env:
         PYO3_USE_ABI3_FORWARD_COMPATIBILITY: "1"
-      run: uv run pytest tests/test_template.py -v --cov=tests --cov-report=term-missing
+      run: uv run pytest tests/test_template.py tests/test_template_lint.py -v --cov=tests --cov-report=term-missing
 
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@671740ac38dd9b0130fbe1cec585b89eea48d3de # v5.5.2
