+++ .gitignore
@@ -25,6 +25,7 @@ __pycache__/
 
 # cache
 .cache/
+.hypothesis/
 .pytest_cache/
 .ruff_cache/
